requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
python-multipart==0.0.9
cohere==5.5.6
pinecone-client==5.0.1
PyPDF2==3.0.1
tqdm==4.66.1
requests==2.32.3
pydantic==2.7.4
streamlit==1.39.0
